+++ src/OpenTelemetry.Instrumentation.AspNet/AspNetInstrumentation.cs
@@ -17,7 +17,7 @@ internal sealed class AspNetInstrumentation : IDisposable
 
     public static readonly Version SemanticConventionsVersion = new(1, 36, 0);
     public static readonly ActivitySource ActivitySource = ActivitySourceFactory.Create<AspNetInstrumentation>(SemanticConventionsVersion);
-    public static readonly Meter Meter = MeterFactory.Create<AspNetInstrumentation>(SemanticConventionsVersion);
+    public static readonly Meter Meter = Metrics.MeterFactory.Create<AspNetInstrumentation>(SemanticConventionsVersion);
 
     public static readonly Histogram<double> HttpServerDuration = Meter.CreateHistogram(
         "http.server.request.duration",