+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
-* Updated OpenTelemetry core component version(s) to `1.15.2`.
-  ([#4080](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4080))
+## 1.15.2
+
+Released 2026-Apr-21
+
+* Updated OpenTelemetry core component version(s) to `1.15.3`.
+  ([#4166](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4166))
 
 * Log error when a metric's dimensions exceed the serialization buffer length.
   ([#4027](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4027))