+++ src/OpenTelemetry.Instrumentation.Remoting/CHANGELOG.md
@@ -2,4 +2,9 @@
 
 ## Unreleased
 
+* This is the first release of `OpenTelemetry.Instrumentation.Remoting` package.
+
 For more details, please refer to the [README](README.md).
+
+* Updated OpenTelemetry core component version(s) to `1.15.3`.
+  ([#4166](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4166))